fix: 에러발생 후 같은 종류 페이지 접속 시 에러바운더리가 유지되는 오류 해결
- 에러바운더리에서 에러가 reset되지 않아 문제 발생
- key를 주어 에러바운더리 reset시키기
