+++ frontend/src/components/notice/NoticeDetailFetcher/index.tsx
@@ -1,13 +1,8 @@
-import { useParams } from 'react-router-dom';
-
 import { useNoticeDetail } from '@hooks';
 
 import NoticeDetail from '../NoticeDetail';
 
-export default function NoticeDetailFetcher() {
-  const { noticeId: NoticeIdParam } = useParams();
-  const noticeId = Number(NoticeIdParam) ?? 0;
-
+export default function NoticeDetailFetcher({ noticeId }: { noticeId: number }) {
   const { data: notice } = useNoticeDetail(noticeId);
 
   if (!notice) return <></>;
