+++ frontend/src/pages/notice/NoticeDetailPage/index.tsx
@@ -1,4 +1,5 @@
 import { Suspense } from 'react';
+import { useParams } from 'react-router-dom';
 
 import ErrorBoundary from '@pages/ErrorBoundary';
 
@@ -9,13 +10,16 @@ import NoticeDetailFetcher from '@components/notice/NoticeDetailFetcher';
 import * as S from './style';
 
 export default function NoticeDetailPage() {
+  const { noticeId: NoticeIdParam } = useParams();
+  const noticeId = Number(NoticeIdParam) ?? 0;
+
   return (
     <Layout isSidebarVisible>
       <S.Container>
         <S.Category tabIndex={0}>VoTogether 공지사항</S.Category>
-        <ErrorBoundary hasIcon={true} hasRetryInteraction={true}>
+        <ErrorBoundary hasIcon={true} hasRetryInteraction={true} key={noticeId}>
           <Suspense fallback={<Skeleton isLarge />}>
-            <NoticeDetailFetcher />
+            <NoticeDetailFetcher noticeId={noticeId} />
           </Suspense>
         </ErrorBoundary>
       </S.Container>
