+++ frontend/src/pages/post/PostDetailPage/index.tsx
@@ -1,4 +1,5 @@
 import { Suspense } from 'react';
+import { useParams } from 'react-router-dom';
 
 import ErrorBoundaryWithNarrowHeader from '@pages/ErrorBoundaryWithNarrowHeader';
 
@@ -9,12 +10,16 @@ import Skeleton from '@components/common/Skeleton';
 import PostDetail from './PostDetail';
 
 export default function PostDetailPage() {
+  const params = useParams() as { postId: string };
+  const postId = Number(params.postId);
+
   return (
     <Layout isSidebarVisible={true} isMobileDefaultHeaderVisible={false}>
       <ErrorBoundaryWithNarrowHeader
         hasIcon={true}
         hasRetryInteraction={true}
         hasHomeInteraction={true}
+        key={postId}
       >
         <Suspense
           fallback={
@@ -23,7 +28,7 @@ export default function PostDetailPage() {
             </MobileLayoutTemplate>
           }
         >
-          <PostDetail />
+          <PostDetail postId={postId} />
         </Suspense>
       </ErrorBoundaryWithNarrowHeader>
     </Layout>
